pyproject.toml
[project]
name = "FunVIP"
version = "0.3.21"
description = "Fungal Validation & Identification Pipeline"
authors = [{name = "Changwan Seo", email = "wan101010@snu.ac.kr"}]
urls = { "Homepage" = "https://github.com/Changwanseo/FunVIP" }
requires-python = ">3.8, <3.13"
license = {text = "GPL-3.0"}
dependencies = [
    "biopython==1.78",
    "ete3==3.1.3",
    "Cython",
    "dendropy",
    "GenMine>=1.0.13",
    "lxml",
    "matplotlib",
    "numpy",
    "openpyxl==3.0.9",
    "pandas==1.4.2",
    "plotly==5.9.0",
    "psutil",
    "pyyaml",
    "sip>=4.19.4",
    "scikit-learn",
    "scipy",
    "tabulate",
    "unidecode==1.2.0",
    "xlrd==2.0.1",
    "xlsxwriter",
    "xmltodict==0.12.0",
    "PyQt5>=5.9.2; sys_platform!='darwin'"
]

[project.scripts]
FunID = "funvip.main:main"
FunVIP = "funvip.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["funvip"]



[build-system]
requires = ["setuptools", "setuptools-ext", "wheel"]
build-backend = "setuptools_ext"

[metadata]
obsoletes-dist = "FunID"

